serverless/requirements.txt
GitPython==3.1.40
pycryptodome==3.18.0
